requirements.txt
alembic==1.11.1
async-timeout==4.0.2
attrs==23.1.0
Authlib==1.2.0
Babel==2.12.1
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
Deprecated==1.2.14
exceptiongroup==1.1.1
flasgger==0.9.7.1
Flask==2.3.2
Flask-BabelEx==0.9.4
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
googleapis-common-protos==1.56.2
grpcio==1.54.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.2
mistune==2.0.5
opentelemetry-api==1.18.0
opentelemetry-exporter-jaeger==1.18.0
opentelemetry-exporter-jaeger-proto-grpc==1.18.0
opentelemetry-exporter-jaeger-thrift==1.18.0
opentelemetry-instrumentation==0.39b0
opentelemetry-instrumentation-flask==0.39b0
opentelemetry-instrumentation-wsgi==0.39b0
opentelemetry-sdk==1.18.0
opentelemetry-semantic-conventions==0.39b0
opentelemetry-util-http==0.39b0
packaging==23.1
passlib==1.7.4
pluggy==1.0.0
protobuf==3.20.3
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==1.10.7
PyJWT==2.7.0
pyrsistent==0.19.3
pytest==7.3.1
python-dotenv==1.0.0
PyYAML==6.0
redis==4.5.5
requests==2.31.0
six==1.16.0
speaklater==1.3
SQLAlchemy==2.0.15
thrift==0.16.0
tomli==2.0.1
typing_extensions==4.6.0
urllib3==2.0.3
Werkzeug==2.3.4
wrapt==1.15.0
WTForms==3.0.1
zipp==3.15.0
